backend/requirements.txt
flask
flask-sqlalchemy
python-dotenv
flask-restful
flask-migrate
flask-jwt-extended
mysql-connector-python
pymysql
python-dotenv